adding back my changes to order_items controller and tests after the hard reset

+++ app/controllers/order_items_controller.rb
@@ -14,47 +14,42 @@ class OrderItemsController < ApplicationController
       session[:order_id] = @order.id
     end
 
+    #if the order is retired then it is not for sale
     @product = Product.find(params[:order_item][:product_id])
+    puts "Product is retired? #{retired?}"
     if retired?
       return
     end
-    #stock logic
-    if params[:order_item][:quantity].to_i > @product.stock.to_i
-      flash.now[:status] = :failure
-      flash.now[:message] = "There is not enough stock. Order a smaller amount"
-      #redirect_to order_items_path
+
+    #if an order_item for the same product already exists in the order, update that order_item by adding the new quantity to that order item.
+    if consoldate_order_items(session[:order_id],params[:order_item][:product_id], params[:order_item][:quantity])
+      redirect_to order_items_path
+      return
+    end
+
+    #if the request is a the first order for a shopping cart, make the orderitem and set the session_id
+    @order_item = OrderItem.new(order_item_params)
+    @order_item.order_id = session[:order_id]
+    if @order_item.save
+      flash[:status] = :success
+      flash[:message] = "Successfully added #{@order_item.product.name} to your cart"
+      redirect_to order_items_path
     else
-
-      if consoldate_order_items(session[:order_id],params[:order_item][:product_id], params[:order_item][:quantity])
-        redirect_to order_items_path
-        return
-      end
-
-      @order_item = OrderItem.new(order_item_params)
-      @order_item.order_id = session[:order_id]
-      if @order_item.save
-        flash[:status] = :success
-        flash[:message] = "Successfully added #{@order_item.product.name} to your cart"
-        redirect_to order_items_path
-      else
-        flash[:status] = :failure
-        flash[:message] = "Could not add #{@order_item.product.name} to your cart"
-        redirect_to root_path
-      end
+      flash[:status] = :failure
+      flash[:message] = "Could not add #{@order_item.product.name} to your cart"
+      redirect_to root_path
     end
   end
 
+
   def update
     @order_item = OrderItem.find(params[:id])
-
-    @product = Product.find(@order_item.product_id)
-
-    if params[:order_item][:quantity].to_i > @product.stock.to_i
-      flash[:status] = :failure
-      flash[:message] = "There is not enough stock. Order a smaller amount"
-      redirect_to order_items_path
-    end
-
+    # @product = Product.find(@order_item.product_id)
+    # if params[:order_item][:quantity].to_i > @product.stock.to_i
+    #   flash[:status] = :failure
+    #   flash[:message] = "There is not enough stock. Order a smaller amount"
+    #   redirect_to order_items_path
+    # end
     @order_item.update_attributes(order_item_params)
     if @order_item.save
       flash[:status] = :success
@@ -69,20 +64,22 @@ class OrderItemsController < ApplicationController
   end
 
   def destroy
-    @order_item = OrderItem.find(params[:id])
-    @product = Product.find(@order_item.product_id)
-    @product.stock += @order_item.quantity
-    @product.save
-    if @order_item.destroy
-      flash[:status] = :success
-      flash[:message] = "Successfully removed #{@order_item.product.name} from your cart"
-      redirect_to order_items_path
-    else
-      flash[:status] = :failure
-      flash[:message] = "Problem encountered when attempting to remove #{@order_item.product.name} from your cart"
-      redirect_to order_items_path
+    #Normal case: if there is an order delete the order item
+    if @order_item = OrderItem.find_by(id: params[:id])
+      if @order_item.destroy
+        flash[:status] = :success
+        flash[:message] = "Successfully removed #{@order_item.product.name} from your cart"
+        redirect_to order_items_path
+        return
+      end
     end
+
+    #if the orderitem does not exists, or the delete fails for some other reason, return that you can not delete the item
+    flash[:status] = :failure
+    flash[:message] = "Problem encountered when attempting to remove this product from your cart"
+    redirect_to order_items_path
   end
+
 
   private
